More simplification of winforms libs.
