+++ src/winforms/toga_winforms/libs/winforms.py
@@ -4,7 +4,7 @@ import clr
 
 clr.AddReference("System.Windows.Forms")
 
-import System.Windows.Forms as WinForms  # noqa: E402
+import System.Windows.Forms as WinForms  # noqa: E402, F401
 from System import Action  # noqa: E402, F401
 from System import Single  # noqa: E402, F401
 from System import Convert  # noqa: E402, F401
@@ -20,16 +20,6 @@ from System.Drawing import FontFamily, FontStyle, SystemFonts  # noqa: E402, F40
 from System.Drawing import Text, Color, Bitmap  # noqa: E402, F401
 from System.Threading.Tasks import Task  # noqa: E402, F401
 
-from toga.constants import LEFT, RIGHT, CENTER, JUSTIFY  # noqa: E402
-from toga.fonts import (
-    MESSAGE,
-    SYSTEM,
-    SERIF,
-    SANS_SERIF,
-    CURSIVE,
-    FANTASY,
-    MONOSPACE,
-)  # noqa: E402
 
 user32 = ctypes.windll.user32
 # shcore dll not exist on some Windows versions
@@ -39,54 +29,3 @@ try:
 except OSError:
     shcore = None
 win_version = Environment.OSVersion.Version
-
-
-def TextAlignment(value):
-    return {
-        LEFT: ContentAlignment.MiddleLeft,
-        RIGHT: ContentAlignment.MiddleRight,
-        CENTER: ContentAlignment.MiddleCenter,
-        JUSTIFY: ContentAlignment.MiddleLeft,
-    }[value]
-
-
-# Justify simply sets Left alignment. Is this the best option?
-def HorizontalTextAlignment(value):
-    return {
-        LEFT: WinForms.HorizontalAlignment.Left,
-        RIGHT: WinForms.HorizontalAlignment.Right,
-        CENTER: WinForms.HorizontalAlignment.Center,
-        JUSTIFY: WinForms.HorizontalAlignment.Left,
-    }[value]
-
-
-def add_handler(cmd):
-    action = cmd.action
-
-    def handler(sender, event):
-        return action(None)
-
-    return handler
-
-
-def win_font_family(value):
-    win_families = {
-        SYSTEM: SystemFonts.DefaultFont.FontFamily,
-        MESSAGE: SystemFonts.MenuFont.FontFamily,
-        SERIF: FontFamily.GenericSerif,
-        SANS_SERIF: FontFamily.GenericSansSerif,
-        CURSIVE: FontFamily("Comic Sans MS"),
-        FANTASY: FontFamily("Impact"),
-        MONOSPACE: FontFamily.GenericMonospace,
-    }
-    for key in win_families:
-        if value in key:
-            return win_families[key]
-    if value in Text.InstalledFontCollection().Families:
-        return FontFamily(value)
-    else:
-        print(
-            "Unable to load font-family '{}', loading {} instead".format(
-                value, SystemFonts.DefaultFont.FontFamily)
-        )
-        return SystemFonts.DefaultFont.FontFamily
